+++ src/List/index.tsx
@@ -29,8 +29,32 @@ const DataEditorList: React.FC = (props: any) => {
                   const displayValue = record[modelField.key];
                   return (
                     <td key={modelIndex}>
-                      {(modelField.type === DataEditorModelFieldType.String &&
-                        displayValue) ||
+                      {(modelField.type === DataEditorModelFieldType.Object && (
+                        <dl>
+                          {modelField.model?.map(
+                            (
+                              childModelField: DataEditorModelField,
+                              childIndex: number
+                            ) => {
+                              return (
+                                <React.Fragment key={childIndex}>
+                                  <dt>
+                                    {childModelField.label
+                                      ? childModelField.label
+                                      : childModelField.key}
+                                  </dt>
+                                  <dd>
+                                    {displayValue &&
+                                      displayValue[childModelField.key]}
+                                  </dd>
+                                </React.Fragment>
+                              );
+                            }
+                          )}
+                        </dl>
+                      )) ||
+                        (modelField.type === DataEditorModelFieldType.String &&
+                          displayValue) ||
                         (modelField.type === DataEditorModelFieldType.Array && (
                           <React.Fragment>
                             {(Array.isArray(displayValue) &&
@@ -55,7 +79,11 @@ const DataEditorList: React.FC = (props: any) => {
                 <ButtonComponent
                   onClick={(e: any) => {
                     e.preventDefault();
-                    if (props.deleteCallback) props.deleteCallback(recordIndex);
+                    if (
+                      props.deleteCallback &&
+                      window.confirm('Do you want to delete this record?')
+                    )
+                      props.deleteCallback(recordIndex);
                   }}
                 >
                   Delete